Resetting Candy to stock CSS and JS
To beat the "booty call" bug, whereby starting a private chat makes
the recipient believe that they've been taken into a private room
(when in fact they're still in the group room), we're resetting all
of the Candy static assets to the stock resources.

Override user agent CSS default on message input

The default Candy CSS doesn't do a very good CSS reset, so in this
case, the user agent stylesheet put a margin in the message form,
pushing the input box up a bit. Set the margin to 0 so that things
look normal.